pyproject.toml
[project]
name = "ml709-adversarial-agents"
version = "0.1.0"
description = "Research project: Designing Robust Agentic AI Workflows under Adversarial Tool Use"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "ML-709 Research Team" }
]
keywords = ["llm", "agents", "adversarial", "robustness", "security"]

dependencies = [
    # LLM - Primary: vLLM with local models
    "vllm>=0.6.0",
    "transformers>=4.40.0",
    "openai>=1.0.0",

    # Core Framework
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=24.0.0",

    # Async
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",

    # Analysis
    "networkx>=3.2.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",

    # Storage
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",

    # Benchmarks
    "datasets>=2.18.0",

    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
cloud = [
    "anthropic>=0.25.0",
    "litellm>=1.40.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

[project.scripts]
run-experiment = "scripts.run_experiment:main"
run-adrs = "scripts.run_adrs_loop:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
]
